requirements.txt
PyYAML==5.3.1
graphviz==0.19.1
Tensorforce==0.6.5
tensorflow==2.6.0
